multi_container_app_Assignment4/app/requirements.txt
# requirements.txt
Flask>=2.2
Flask-SQLAlchemy==3.0.0
Flask-Caching==1.10.1
psycopg2-binary==2.9.1
redis==4.1.1
